requirements.txt
yandex-cloud-ml-sdk==0.2.4
psycopg2-binary==2.9.10
python-dotenv==1.0.1
